# main/docs/swagger.yaml
basePath: /v1
definitions:
  handler.httpError:
    properties:
      message:
        example: Some error message
        type: string
    type: object
...
  request.RegisterReq:
    properties:
      email:
        example: example@economicus.kr
        type: string
      nickname:
        example: user nickname
        type: string
      password:
        example: some password
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Economicus 메인 서버
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.economicus.kr/
  title: Economicus Main Server
  version: 1.0.0
paths:
...
  /register:
    post:
      consumes:
      - application/json
      description: 이메일, 비밀번호로 유저 회원가입
      parameters:
      - description: A user information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Register a user
      tags:
      - user
  /users:
    get:
      consumes:
      - application/json
      description: 이메일, 비밀번호로 유저 회원가입
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: A user information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: ""
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/handler.httpError'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/handler.httpError'
        "404":
          description: Not found error
          schema:
            $ref: '#/definitions/handler.httpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.httpError'
      summary: Return all users
      tags:
      - user
schemes:
- https
- http
securityDefinitions:
  JWT:
    type: basic
swagger: "2.0"